rename JS file in lowercase for clarity
