+++ src/frontend/scripts/components/chess/chessboardElements.js
